+++ test_suite/tests/integer.rs
@@ -500,7 +500,7 @@ mod traits {
         assert_eq!(age.to_string(), "35");
     }
 
-    #[cfg(feature = "serde1")]
+    #[cfg(feature = "serde")]
     #[test]
     fn test_trait_serialize() {
         #[nutype]
@@ -512,7 +512,7 @@ mod traits {
         assert_eq!(offset_json, "-280");
     }
 
-    #[cfg(feature = "serde1")]
+    #[cfg(feature = "serde")]
     #[test]
     fn test_trait_deserialize_without_validation() {
         #[nutype]
@@ -530,7 +530,7 @@ mod traits {
         }
     }
 
-    #[cfg(feature = "serde1")]
+    #[cfg(feature = "serde")]
     #[test]
     fn test_trait_deserialize_with_validation() {
         #[nutype(validate(min = 13))]