+++ CHANGELOG.md
@@ -2,9 +2,10 @@
 ### v0.2.0 - 2023-??-??
 
 * [BREAKING] Rename string validator `present` -> `not_empty`. Rename error variant `Missing` -> `Empty`.
+* [BREAKING] Rename feature `serde1` to `serde`.
 * Introduce `new_unchecked` feature flag, that allows to bypass sanitization and validation.
 * Support derive of `JsonSchema` of `schemars` crate (requires `schemars08` feature).
-* Support string validation with `regex` (requires `regex1` feature).
+* Support string validation with `regex` (requires `regex` feature).
 
 ### v0.1.1 - 2023-02-11
 * Initial release